War of Weapons Board Game
In the sixth project of my OpenClassrooms front-end web development course I had to create a board game in which players could move around a board, one to three spaces with each move, pick up and drop weapons to reflect the damage they could inflict on another player, and ultimately fight the opponent by either attacking or defending.  Each attacks reduces the opponent's score by the amount of their weapon's damage, which depends on the quality of their weapons. If a player chooses to defend then the damage inflicted by the opponent is halved.  The first player to have their score reduced to zero loses. The project requires the use of vanilla JavaScript, jQuery, object-oriented JavaScript, DOM manipulation, CSS, bootstrap, and html.